# Conflicts:
#	src/main/resources/data.sql

+++ src/main/java/com/rcodingschool/carrepair/Domain/Repair.java
@@ -3,7 +3,6 @@ package com.rcodingschool.carrepair.Domain;
 
 import javax.persistence.*;
 import java.io.Serializable;
-
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -25,7 +24,7 @@ public class Repair implements Serializable {
     private String repairTasks;
 
     @Column(name = "RepairTotalCost", nullable = false)
-    private Integer repairTotalCost;
+    private Long repairTotalCost;
 
     @Column(name = "RepairTypeID", nullable = false)
     private Short repairTypeID;
@@ -80,11 +79,11 @@ public class Repair implements Serializable {
         this.repairTasks = repairTasks;
     }
 
-    public Integer getRepairTotalCost() {
+    public Long getRepairTotalCost() {
         return repairTotalCost;
     }
 
-    public void setRepairTotalCost(Integer repairTotalCost) {
+    public void setRepairTotalCost(Long repairTotalCost) {
         this.repairTotalCost = repairTotalCost;
     }
